Handle connection termination messages via an error on the receive messages
This was a difficult design decision. We could continue to return an empty array, and expect users to understand what was going on. This has already proven to be difficult on a few projects so I decided to explicitly throw an error to the caller of receive. This indicates to the caller that no data could be read because the connection was signaled closed, and helps inform people that sockets don't send empty data. An error also clearly indicates that this Socket instance can no longer be used. These errors should be handled as less severe, but the shutdown process is the same as other errors.